backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
langchain==0.0.267
langchain-community==0.0.16
langchain-core>=0.1.16,<0.2.0
langchain-openai==0.0.2
openai>=1.6.1
pandas==2.0.3
openpyxl==3.1.2
matplotlib==3.7.2
python-dotenv==1.0.0
pydantic==2.3.0
aiofiles==23.2.1
jinja2==3.1.2
uuid==1.30
asyncio==3.4.3
numpy>=1.24.0
typing-extensions>=4.5.0
gunicorn==21.2.0 